Move the object model to siameseX model package
