--- .idea/libraries/Maven__com_fasterxml_jackson_core_jackson_databind_2_10_0.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: com.fasterxml.jackson.core:jackson-databind:2.10.0">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/com/fasterxml/jackson/core/jackson-databind/2.10.0/jackson-databind-2.10.0.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/com/fasterxml/jackson/core/jackson-databind/2.10.0/jackson-databind-2.10.0-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/com/fasterxml/jackson/core/jackson-databind/2.10.0/jackson-databind-2.10.0-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
