+++ webapp/src/main/java/com/siamesex/standalone/model/Commit.java
@@ -1,4 +1,4 @@
-package crest.siamese.githubUtils;
+package com.siamesex.standalone.model;
 
 import org.json.simple.JSONObject;
