+++ webapp/src/main/java/com/siamesex/standalone/model/Hunk.java
@@ -1,4 +1,4 @@
-package crest.siamese.githubUtils;
+package com.siamesex.standalone.model;
 
 public interface Hunk {
