+++ webapp/src/main/java/com/siamesex/standalone/model/HunkQuery.java
@@ -1,4 +1,4 @@
-package crest.siamese.githubUtils;
+package com.siamesex.standalone.model;
 
 public class HunkQuery implements Hunk {
